% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddRes2Userm.R
\name{AddRes2Userm}
\alias{AddRes2Userm}
\title{Add a Res object to a Userm object}
\usage{
AddRes2Userm(Res, Userm)
}
\arguments{
\item{Res}{A Res object.}

\item{Userm}{A Userm object, generated with CreateUserm function.}
}
\value{
The updated \code{Userm} object with the new \code{Res} object added.
}
\description{
This function validates and adds a Res object (\code{Res}) to a Userm object (\code{Userm}). It checks whether the \code{Res} object has a valid ID, matching detectors, and properly structured matrices (\code{interceptMtx} and \code{slopMtx}). If all checks pass, the Res object is added to the \code{Userm$Res} list.
}
\examples{
\dontrun{
Res <- CreateRes(id, R, A)
Userm <- CreateUserm(A)
Userm <- AddRes2Userm(Res, Userm)
}
}
