% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRes_covScatter.R
\name{checkRes_covScatter}
\alias{checkRes_covScatter}
\title{Plot Covariance Scatter Between Two Detectors with Regression Line}
\usage{
checkRes_covScatter(Res, detector1, detector2)
}
\arguments{
\item{Res}{A list containing the following components:
\itemize{
\item \code{bin_mids}: A numeric vector of bin midpoints.
\item \code{detectors}: A character vector of detector names.
\item \code{cov_matrices}: A 3D array of covariance values indexed by detector pairs and bins.
\item \code{slopMtx}: A matrix of slope values for each detector pair.
\item \code{interceptMtx}: A matrix of intercept values for each detector pair.
}}

\item{detector1}{Name of the one detector (character string).}

\item{detector2}{Name of the another detector (character string).}
}
\value{
A \code{ggplot} object showing the scatter plot of covariance values and a regression line.
}
\description{
This function generates a scatter plot of covariance values between two specified detectors
across bin midpoints, using data from a result object \code{Res}. A regression line is added
using the corresponding slope and intercept values, which are also displayed in the plot title.
}
\examples{
\dontrun{
checkRes_covScatter(Res = ResObj,
detector1 = ResObj$detectors[1],
detector2 = ResObj$detectors[2])
}
}
