% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRes_slopMtx.R
\name{checkRes_slopMtx}
\alias{checkRes_slopMtx}
\title{Visualize the Slope Matrix from a Result Object}
\usage{
checkRes_slopMtx(
  Res,
  min = -1,
  mid = 0,
  max = 1,
  mincolor = "blue",
  midcolor = "white",
  maxcolor = "red"
)
}
\arguments{
\item{Res}{A Res object.}

\item{min}{Numeric value representing the minimum of the color scale. Default is -1.}

\item{mid}{Numeric value representing the midpoint of the color scale. Default is 0.}

\item{max}{Numeric value representing the maximum of the color scale. Default is 1.}

\item{mincolor}{Color corresponding to the minimum value. Default is "blue".}

\item{midcolor}{Color corresponding to the midpoint value. Default is "white".}

\item{maxcolor}{Color corresponding to the maximum value. Default is "red".}
}
\value{
A \code{Heatmap} object from the ComplexHeatmap package, visualizing the \code{slopMtx} with annotated values.
}
\description{
This function generates a heatmap visualization of the \code{slopMtx} matrix contained within a result object \code{Res}.
It uses a customizable color gradient and overlays numeric values on each cell.
}
\examples{
\dontrun{
  checkRes_slopMtx(Res)
}

}
