% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRes.R
\name{CreateRes}
\alias{CreateRes}
\title{Create a Res object}
\usage{
CreateRes(id, R, A)
}
\arguments{
\item{id}{A character string representing the identifier for the result object. If missing, defaults to \code{"x"}.}

\item{R}{A numeric matrix of raw signal data (cells × detectors), typically extracted from an FCS file.}

\item{A}{A numeric unmixing matrix (detectors × fluors), typically extracted from a single-stained control (SCC) FCS file.}
}
\value{
A list representing the initialized residual object, containing the input matrices and placeholders for analysis results.
}
\description{
This function creates a structured result object (\code{Res}) for storing residual model information extracted from spectral flow scc file.
It validates the input raw signal matrix \code{R} and the unmixing matrix \code{A}, ensuring that their dimensions
and identifiers are compatible. If valid, it returns a list containing initialized fields for further analysis.
}
\examples{
\dontrun{
R <- matrix(runif(100), nrow = 10, ncol = 5)
colnames(R) <- paste0("D", 1:5)
A <- matrix(runif(25), nrow = 5, ncol = 5)
rownames(A) <- paste0("D", 1:5)
colnames(A) <- paste0("F", 1:5)
Res <- CreateRes("Sample1", R, A)
}
}
