% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigExtract.R
\name{ExtractSig}
\alias{ExtractSig}
\title{Extract a normalized signature from a single-stained control (SCC)}
\usage{
ExtractSig(
  df_pos,
  df_neg,
  cols,
  method,
  PrimaryName,
  SecondaryName,
  id,
  instrument
)
}
\arguments{
\item{df_pos}{A data frame containing positive control signal data (cells x detectors).}

\item{df_neg}{A data frame containing negative control signal data (cells x detectors).}

\item{cols}{A character vector specifying the column names (detector channels) to be used for signature extraction.}

\item{method}{A function used to summarize each column (e.g., \code{median}, \code{mean}).}

\item{PrimaryName}{A character string specifying the primary name of the fluorophore.}

\item{SecondaryName}{A character string specifying the secondary name of the fluorophore or marker.}

\item{id}{A character string used to identify the signature object.}

\item{instrument}{A character string used to specify the instrument used to generate the raw data, e.g. Aurora5L, Xenith.}
}
\value{
A list containing:
\item{id}{The identifier string.}
\item{PrimaryName}{The primary name of the fluorophore.}
\item{SecondaryName}{The secondary name of the fluorophore or marker.}
\item{Signature}{A normalized numeric vector representing the extracted signature.}
}
\description{
This function computes a normalized signature vector by comparing positive and negative control data
for a set of specified detector channels. It applies a summary method (e.g., \code{median}, \code{mean})
to each column in the positive and negative data frames, calculates the difference, and normalizes the result.
}
\examples{
\dontrun{
df_pos <- data.frame(D1 = rnorm(100, 100), D2 = rnorm(100, 200))
df_neg <- data.frame(D1 = rnorm(100, 10), D2 = rnorm(100, 20))
sig <- ExtractSig(df_pos, df_neg, cols = c("D1", "D2"), method = mean,
                  PrimaryName = "FITC", SecondaryName = "CD3", id = "FITC_CD3")
}
}
