% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColor.R
\name{getColor}
\alias{getColor}
\title{Interpolate a color based on a numeric value}
\usage{
getColor(val, min, mid, max, colormin, colormid, colormax)
}
\arguments{
\item{val}{A numeric value to be mapped to a color.}

\item{min}{The minimum value of the range.}

\item{mid}{The midpoint value of the range.}

\item{max}{The maximum value of the range.}

\item{colormin}{The color (in HEX format) corresponding to the minimum value.}

\item{colormid}{The color (in HEX format) corresponding to the midpoint value.}

\item{colormax}{The color (in HEX format) corresponding to the maximum value.}
}
\value{
A HEX color string representing the interpolated color.
}
\description{
This function interpolates a color between three specified colors (minimum, midpoint, and maximum)
based on the position of a numeric value within a defined range.
}
\examples{
getColor(0.75, 0, 0.5, 1, "#0000FF", "#00FF00", "#FF0000")
}
