% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRes.R
\name{getRes}
\alias{getRes}
\title{Retrieve Result Object by ID}
\usage{
getRes(id, custom_dir = NULL)
}
\arguments{
\item{id}{A character string representing the ID of the result object to retrieve.}
}
\value{
An R object containing the result data associated with the specified ID.
}
\description{
This function loads a serialized result object from the internal package directory
based on a given ID. The result object is expected to be stored as an RDS file
named \code{ResObj_<id>.rds} within the \code{res} folder of the \code{USERM} package.
}
\details{
The function constructs the file path using \code{system.file()} and checks whether
the corresponding RDS file exists. If the file is found, it is read and returned.
If not, the function throws an error indicating that the result object could not be found.
}
\examples{
\dontrun{
  res <- getRes("SCC_Bead_CD4_NFR700")
  print(res)
}

}
