% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopEstimation.R
\name{SlopEstimation}
\alias{SlopEstimation}
\title{Estimate slope and intercept matrices from residuals}
\usage{
SlopEstimation(
  Res,
  count_thre,
  bin_num = 30,
  bin_method = "percentile",
  z_thre = 3,
  ...
)
}
\arguments{
\item{Res}{A result object created by \code{\link{CreateRes}}, containing raw signal matrix \code{R}, unmixing matrix \code{A}, and detector information.}

\item{count_thre}{An integer specifying the minimum number of data points required in a bin to perform covariance estimation.}

\item{bin_num}{An integer specifying the number of bins to divide the data into. Default is 30.}

\item{bin_method}{A character string specifying the binning method. Must be either \code{"absolute"} or \code{"percentile"}. Default is \code{"percentile"}.}

\item{z_thre}{A numeric threshold for removing outliers based on z-scores of the first column of matrix \code{B}. Default is 3.}

\item{...}{Additional arguments passed to internal functions (currently unused).}
}
\value{
The updated \code{Res} object with the following fields added:
\item{bin_mids}{Midpoints of each bin used for slope estimation.}
\item{bin_counts}{Number of data points in each bin.}
\item{cov_matrices}{Robust covariance matrices computed for each bin.}
\item{interceptMtx}{Matrix of intercepts estimated from robust linear models.}
\item{slopMtx}{Matrix of slopes estimated from robust linear models.}
\item{par}{A list containing binning parameters used in the estimation.}
}
\description{
This function performs robust estimation of slope and intercept matrices based on residuals from a signal reconstruction model.
It bins the data using either absolute or percentile-based methods, computes robust covariance matrices for each bin,
and fits robust linear models to estimate the relationship between covariance values and bin midpoints.
}
\examples{
\dontrun{
Res <- CreateRes("F1", R, A)
Res <- SlopEstimation(Res, count_thre = 30, bin_num = 30, bin_method = "percentile")
}
}
